requirements.txt
discord.py>=2.0.0
sqlite3
asyncio>=3.4.3
